miRScore V0.3
Precision is now calculated on an individual library basis. A new column was added to results file 'n_libraries' which reports how many libraries contained reads for the miRNA duplex with high enough precision. 

A new feature was added that predicts the coordinates for miRNAs which failed scoring, reanalyzes these predicted loci, and outputs results to a separate file.